Enable encoding & sending large data to server
- now using XtLarge/DtLarge and 56-bits for length in headers to encode
large data (> 16MB)
- added helpful comments about QAP protocol
- when encoding SexpList, use Sexp.Value instead of Sexp Enumerator -
Sexp.Value is more straightforward
- bug fix: string data in SubmitCommand that, when encoded, generated
bytes whose length was perfectly divisible by 4, was not null
terminated.  Causes server to crash when assigning to symbol who's
length is divisible by 4 in bytes
- SubmitCommand now supports int data which is used with commands such
as CMD_setBufferSize
- Message header now uses bytes 12 thru 16 for bits 32-63 of the message
length, instead of hard-coded to 0
- added xunit.extensions for more unit test capabilities
- added Qap1Test to test large data feature
- Rservice now supports optional parameter maxInputBufferSizeInKb which
enables the server to start with a different input buffer size.  Ensures
that we can test sending large data to the server
- Rservice sleeps for 3 seconds after launching the server and before
connecting the client to provide enough time for the server to startup
-